refactor(shared): align interfaces with domain modules
