+++ packages/shared/src/automation/index.ts
