--- packages/shared/src/subpaths/session/events.ts
@@ -1,15 +1,0 @@
-export { clientMessageSchema } from "../../types/websocket";
-export type { ClientMessage } from "../../types/websocket";
-
-export { serverMessageSchema } from "../../types/server-messages";
-export type { ServerMessage } from "../../types/server-messages";
-
-export { sandboxEventSchema } from "../../types/sandbox-events";
-export type {
-  AgentEvent,
-  EventResponse,
-  ListEventsResponse,
-  SandboxEvent,
-} from "../../types/sandbox-events";
-
-export type { EventType, GitSyncStatus } from "../../types/statuses";
